requirements.txt
cffi==1.9.1
cryptography==1.7.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.17
lxml==3.7.1
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
PySocks==1.6.5
python-libtorrent==1.2.0
requests==2.12.4
six==1.10.0
subscope==1.0.2
